pyproject.toml
[project]
name = "datasetgenerator"
version = "0.1.0"
description = "Enhance, scale and generate captions for images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"torch==2.7.1",
	"transformers==4.57.6",
	"diffusers",
	"pillow",
	"accelerate",
	"einops",
	"timm",
	"datasets",
	"tqdm",
	"av",
	"ffmpeg-python",
	"opencv-contrib-python",
	"ruff",
]

[[tool.uv.index]]
name="pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128" 
default = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.uv.sources]
# Tell uv to look for torch, on the custom index
torch = [ { index = "pytorch-cu128" } ]
